python/requirements.txt
PySide6>=6.5.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
